agregado método para restringir la eliminación de registros si estos están relacionados con otros modelos que contengan información

+++ modules/Budget/Http/Controllers/BudgetAccountController.php
@@ -185,10 +185,12 @@ class BudgetAccountController extends Controller
      */
     public function destroy($id)
     {
-        dd($id);
         $budgetAccount = BudgetAccount::find($id);
 
         if ($budgetAccount) {
+            if ($budgetAccount->restrictDelete()) {
+                return response()->json(['error' => true, 'message' => 'El registro no se puede eliminar'], 200);
+            }
             $budgetAccount->delete();
         }
